added one mapping for creating a product

+++ Looms.PoS/Looms.PoS.Application/Models/Requests/Product/VariationRequest.cs
@@ -2,5 +2,5 @@ namespace Looms.PoS.Application.Models.Requests.Product;
 
 public record VariationRequest(String Name){
     public String Name = Name;
-    public float? Price;
+    public decimal? Price;
 }
